LayoutTests/imported/w3c/web-platform-tests/websockets/constructor/011-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:49001/handshake_protocol' failed: Error during WebSocket handshake: Sec-WebSocket-Protocol mismatch

PASS WebSockets: protocol mismatch 
